amath_images/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-cloud-storage
cairosvg
pydantic